+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.0.0] - 2023-01-15
 ### Changed
 - Increase Unity minimum version to 2021.1
 - AbstractState OnEnter, OnUpdate and OnExit are renamed to OnEntered, OnUpdated and OnExited
@@ -44,7 +45,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/AnimatorStates/compare/1.2.0...main
+[Unreleased]: https://github.com/HyagoOliveira/AnimatorStates/compare/2.0.0...main
+[2.0.0]: https://github.com/HyagoOliveira/AnimatorStates/tree/2.0.0/
 [1.2.0]: https://github.com/HyagoOliveira/AnimatorStates/tree/1.2.0/
 [1.1.0]: https://github.com/HyagoOliveira/AnimatorStates/tree/1.1.0/
 [1.0.0]: https://github.com/HyagoOliveira/AnimatorStates/tree/1.0.0/